requirements.txt
agno
openai
python-dotenv
gradio
yfinance
ddgs

